Handle destroy failures in more SJR-friendly way
SJR requires `render` instead of `redirect_to`.  The "success" behavior
of `destroy` can easily be overridden, but not the "error" behavior.
Thus change the "error" behavior to render, a la `create` and `update`.

This behavior is arguably better for non-SJR reponses as well, because
it deterministically calls attention to the failed-to-destroy resource.