Test scaffold generator with namespaced resource
This test is combined with the "skip conflicting locales" test (as
the non-namespaced resource and "generate missing locales" tests are)
to minimize the number of times the Rails scaffold generator must be
invoked.  The Rails scaffold generator is very slow, and dominates the
current test suite execution time, even with these tests combined.