Optimize generators and refactor generator tests
Invoking auxiliary generators via `Rails::Generators::Base#generate`
shells out to the `rails` command, which is very slow and which requires
certain files to be already present in the destination directory.  By
overriding `#generate` with an implementation that does not shell out,
we avoid both these problems.

The resultant performance boost affords room to de-parallelize the
generator tests (except where isolation is needed) for simplicity, and
to split up compound tests for clarity.

Here are some test suite timings from the development machine:

* Previous (compound tests, parallelized): 29 seconds
* Refactored (split tests, not parallelized): 57 seconds
* Optimized (split tests, not parallelized): 3.3 seconds